ODP-000: Moved all odp node module code to respective modules.
