+++ modules/custom/odp_base/modules/odp_blocks/odp_blocks.info.yml
@@ -11,4 +11,4 @@ dependencies:
   - allowed_formats:allowed_formats
   - drupal:media_library
   - textfield_counter:textfield_counter
-  - odp_node:odp_node
+  - odp_product:odp_product
