+++ modules/custom/odp_base/modules/odp_breadcrumb/odp_breadcrumb.info.yml
@@ -7,3 +7,4 @@ core_version_requirement: ^8 || ^9
 
 dependencies:
   - odp_base:odp_base
+  - odp_user:odp_user
