--- modules/custom/odp_node/odp_node.info.yml
@@ -1,17 +1,0 @@
-name: Developer Portal Node
-description : 'Extend the nodes and content specific for developer portal.'
-type: module
-package: ODP
-core: 8.x
-core_version_requirement: ^8 || ^9
-project: odp_node
-
-dependencies:
-  - allowed_formats:allowed_formats
-  - drupal:media_library
-  - textfield_counter:textfield_counter
-  - drupal:layout_builder
-  - drupal:content_moderation
-  - hook_event_dispatcher:hook_event_dispatcher
-  - odp_user:odp_user
-  - group:group
