--- modules/custom/odp_node/odp_node.services.yml
@@ -1,28 +1,0 @@
-services:
-  odp_node.node:
-    class: Drupal\odp_node\Node
-    arguments: ['@database', '@current_user', '@entity_type.manager']
-
-  odp_node.form_alter_subscriber:
-    class: Drupal\odp_node\EventSubscriber\FormAlterSubscriber
-    arguments: ['@request_stack', '@entity_type.manager', '@odp_user.organisation']
-    tags:
-      - { name: event_subscriber }
-
-  odp_node.entity_add_subscriber:
-    class: Drupal\odp_node\EventSubscriber\EntityAddSubscriber
-    tags:
-      - { name: event_subscriber }
-
-  odp_node.usergroup:
-    class: Drupal\odp_node\UserGroupService
-    arguments: ['@current_user', '@entity_type.manager', '@group.membership_loader']
-
-  odp_node.access_check:
-    class: Drupal\odp_node\EventSubscriber\Access
-    tags:
-      - { name: event_subscriber }
-
-  odp_node.program_helper:
-    class: Drupal\odp_node\ProgramHelper
-    arguments: ['@database', '@entity_type.manager']
