+++ modules/custom/odp_product/odp_product.info.yml
@@ -4,3 +4,6 @@ type: module
 package: ODP
 core: 8.x
 core_version_requirement: ^8 || ^9
+
+dependencies:
+  - group:group
