+++ modules/custom/odp_program/odp_program.info.yml
@@ -4,3 +4,7 @@ type: module
 package: ODP
 core: 8.x
 core_version_requirement: ^8 || ^9
+
+dependencies:
+  - hook_event_dispatcher:hook_event_dispatcher
+  - group:group
