+++ modules/custom/odp_program/src/EventSubscriber/EntityAddSubscriber.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\odp_node\EventSubscriber;
+namespace Drupal\odp_program\EventSubscriber;
 
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
