+++ modules/custom/odp_user/modules/odp_login_redirect/odp_login_redirect.links.menu.yml
@@ -2,10 +2,10 @@ odp_login_redirect.login_redirect_settings:
   title: 'Login Redirect Settings'
   route_name: odp_login_redirect.login_redirect_settings
   description: 'Redirect user as per role menu after login'
-  parent: odp_node.admin_index
+  parent: odp_base.admin_index
 
 odp_login_redirect.logout_redirect_settings:
   title: 'Logout Redirect Settings'
   route_name: odp_login_redirect.logout_redirect_settings
   description: 'Redirect user as per role menu after logout'
-  parent: odp_node.admin_index
+  parent: odp_base.admin_index
