+++ modules/custom/odp_user/odp_user.info.yml
@@ -4,5 +4,7 @@ type: module
 package: ODP
 core: 8.x
 core_version_requirement: ^8 || ^9
+
 dependencies:
   - odp_domain:odp_domain
+  - odp_program:odp_program
